langchain_chatbot/requirements.txt
langchain==0.2.1
langchain-community==0.2.1
langchain-cli==0.0.23
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
chromadb==0.5.0
langchain-chroma==0.1.1
pdfplumber==0.11.0
sentence-transformers==3.0.0
FlagEmbedding==1.2.10
torch --index-url https://download.pytorch.org/whl/cu118 #be careful with cuda dependencies. If this is a problem, use CPU but could be too slow
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118
bitsandbytes==0.43.1
accelerate==0.30.1
gradio==4.31.5
gradio_client==0.16.4